chore: clean-up - component-ize this mess - return proper status code (204->200) - use pending-ui updates (a.k.a spinners) - add `.env.example` - use enums in the zod objects were due - conditionally display available data in the email
