+++ .env.example
@@ -1,0 +1,2 @@
+EMAIL_USER=""
+EMAIL_PASSWORD=""
