+++ src/routes/+layout.svelte
@@ -1,6 +1,9 @@
 <script>
+	import { ModeWatcher } from 'mode-watcher';
 	import '../app.pcss';
 </script>
+
+<ModeWatcher />
 
 <main>
 	<slot />
