+++ src/routes/service-form.svelte
@@ -2,12 +2,12 @@
 	import { Hr, Html, Text } from 'svelte-email';
 	import type { TServiceForm } from './schema';
 
-	export let payload: TServiceForm | undefined;
+	export let payload: TServiceForm;
 </script>
 
 <Html lang="en">
 	<Text>
-		Hello, {payload?.other_names}.
+		Hello, {payload.other_names}.
 	</Text>
 	<Hr />
 	<Text>
@@ -19,79 +19,88 @@
 	<Hr />
 	<Text>
 		<strong>Full Name:</strong>
-		{payload?.other_names}
-		{payload?.surname}
+		{payload.other_names}
+		{payload.surname}
 	</Text>
 	<Text>
 		<strong>Email:</strong>
-		{payload?.email}
+		{payload.email}
 	</Text>
-	<Text>
-		<strong>Phone Number:</strong>
-		{payload?.phone_number}
-	</Text>
-	<Text>
-		<strong>National ID:</strong>
-		{payload?.nid}
-	</Text>
-	<Text>
-		<strong>Passport Number:</strong>
-		{payload?.passport_number}
-	</Text>
+	{#if payload.phone_number}
+		<Text>
+			<strong>Phone Number:</strong>
+			{payload.phone_number}
+		</Text>
+	{/if}
+	{#if payload.nid}
+		<Text>
+			<strong>National ID:</strong>
+			{payload.nid}
+		</Text>
+	{:else if payload.passport_number}
+		<Text>
+			<strong>Passport Number:</strong>
+			{payload.passport_number}
+		</Text>
+	{/if}
 	<Text>
 		<strong>Company Name:</strong>
-		{payload?.company_name}
+		{payload.company_name}
 	</Text>
 	<Text>
 		<strong>Business Type:</strong>
-		{payload?.business_type}
+		{payload.business_type}
 	</Text>
 	<Text>
 		<strong>Business Province:</strong>
-		{payload?.business_province}
+		{payload.business_province}
 	</Text>
 	<Text>
 		<strong>Applicant Province:</strong>
-		{payload?.applicant_province}
+		{payload.applicant_province}
 	</Text>
 	<Text>
 		<strong>Product Name:</strong>
-		{payload?.product_name}
+		{payload.product_name}
 	</Text>
 	<Text>
 		<strong>Product Category:</strong>
-		{payload?.product_category}
+		{payload.product_category}
 	</Text>
 	<Text>
 		<strong>Quantity of Products:</strong>
-		{payload?.quantity_of_products}
+		{payload.quantity_of_products}
 	</Text>
 	<Text>
 		<strong>Unit of Measurement:</strong>
-		{payload?.unit_of_measurement}
+		{payload.unit_of_measurement}
 	</Text>
-	<Text>
-		<strong>Weight (kg):</strong>
-		{payload?.weight_kg}
-	</Text>
+	{#if payload.weight_kg}
+		<Text>
+			<strong>Weight (kg):</strong>
+			{payload.weight_kg}
+		</Text>
+	{/if}
 	<Text>
 		<strong>Description of Products:</strong>
-		{payload?.description_of_products}
+		{payload.description_of_products}
 	</Text>
 	<Text>
 		<strong>Purpose of Importation:</strong>
-		{payload?.purpose_of_importation}
+		{payload.purpose_of_importation}
 	</Text>
-	<Text>
-		<strong>Specify Purpose of Importation:</strong>
-		{payload?.specify_purpose_of_importation}
-	</Text>
+	{#if payload.specify_purpose_of_importation}
+		<Text>
+			<strong>Specify Purpose of Importation:</strong>
+			{payload.specify_purpose_of_importation}
+		</Text>
+	{/if}
 	<Text>
 		<strong>Registration Date:</strong>
-		{payload?.registration_date}
+		{payload.registration_date}
 	</Text>
 	<Text>
 		<strong>TIN Number:</strong>
-		{payload?.tin_number}
+		{payload.tin_number}
 	</Text>
 </Html>
